test: добавлены тесты для FilmController

+++ src/main/java/ru/yandex/practicum/filmorate/model/Film.java
@@ -14,5 +14,5 @@ public class Film {
     private String name;
     private String description;
     private LocalDate releaseDate;
-    private long duration;
+    private int duration;
 }
